【バグ修正】アニメーションが発生しないバグの修正
ライブラリの問題。int/intだったら0か1にしかならないのを忘れていた
新たにバグを発見。攻撃で相手を撃破できると落ちる。対策として攻撃処理をAttackScene内で行う事が考えられる。

+++ ActiveWarSimulation/ActiveWarSimulation/ToolsLib.cpp
@@ -312,7 +312,7 @@ void Easing::SetTarget(int i_endx,bool initflame){
 
 void Easing::Update(){
 	double ft;//‘‰ÁŠ„‡
-	const double fullRate=flame/maxflame;
+	const double fullRate=(double)flame/maxflame;
 	if(!GetEndFlag()){
 		if(maxflame>0){
 			switch(function){
